Adds more download checks, improves output

+++ auto-transcoder.sh
@@ -7,24 +7,24 @@ URL="";
 FILENAME="";
 
 getURL() {
-	echo "Please insert the video's URL: ";
+	printf "\n%s" "Please insert the video's URL: ";
 	read -r URL;
 
 	if [[ -z $URL ]]; then
-		echo "Error: Empty URL provided";
+		printf "\n%s" "Error: Empty URL provided";
 		exit 1;
 	else
-		echo "$URL";
+		printf "\n%s" "$URL";
 	fi;
 }
 
 getFileName() {
 	while [ true ]; do
-		echo "Please insert the media's filename: ";
+		printf "\n%s" "Please insert the media's filename: ";
 		read -r FILENAME;
 
 		if [[ -f "$FILENAME" ]]; then
-			echo "A file already exists, with that name. Please insert a different one.";
+			printf "\n%s" "A file already exists, with that name. Please insert a different one.";
 		else
 			break;
 		fi;
@@ -33,24 +33,24 @@ getFileName() {
 
 while [ "$OPTION" != "q" ]; do
 
-	echo "                ###########							 ";
-	echo "What do you want to do?";
-	echo "1. View available formats of media.";
-	echo "2. Download media.";
-	echo "3. Download & Transcode media to highest quality audio (.ogg).";
-	echo "['q': exit]";
-	echo ": ";
+	printf "\n%s" "                ###########							 ";
+	printf "\n%s" "What do you want to do?";
+	printf "\n%s" "1. View available formats of media.";
+	printf "\n%s" "2. Download media.";
+	printf "\n%s" "3. Download & Transcode media to highest quality audio (.ogg).";
+	printf "\n%s" "['q': exit]";
+	printf "\n%s" ": ";
 	read -r OPTION;
 
 	if [[ "$OPTION" == "q" ]]; then
-		echo "Bye!";
+		printf "\n%s" "Bye!";
 		exit 0;
 	elif [[ "$OPTION" == "1" ]]; then
 		getURL;
 		youtube-dl "$URL" -F;
 	elif [[ "$OPTION" == "2" ]]; then
 		getURL;
-		echo "Please insert the media's format id: ";
+		printf "\n%s" "Please insert the media's format id: ";
 		read -r FORMAT;
 
 		getFileName;
@@ -58,17 +58,27 @@ while [ "$OPTION" != "q" ]; do
 	elif [[ "$OPTION" == "3" ]]; then
 		getURL;
 		ID=`youtube-dl "$URL" --get-id`;
-		TEMP="./$ID.webm";
 
-		getFileName;
-		TARGET="$FILENAME";
+		if [[ $? == 0 ]]; then
+			TEMP="./$ID.webm";
 
-		youtube-dl "$URL" -f 171 --id -q;
-		ffmpeg -loglevel panic -i "$TEMP" -vn -b:a 128k -vol 270 -codec libvorbis "$TARGET";
-		rm "$TEMP";
+			getFileName;
+			TARGET="$FILENAME";
 
-		echo "      ############		 ";
-		echo "Audio ready @:\"./$TARGET\"";
-		echo "      ############		 ";
-	fi;
+			youtube-dl "$URL" -f 171 --id -q;
+
+			if [[ $? == 0 ]]; then
+				ffmpeg -loglevel panic -i "$TEMP" -vn -b:a 128k -vol 270 -codec libvorbis "$TARGET";
+				printf "\n%s" "      ############		 ";
+				printf "\n%s" "Audio ready @:\"./$TARGET\"";
+				printf "\n%s" "      ############		 ";
+			else
+				printf "\n%s" "Something went wrong while downloding the video.";
+			fi
+
+			rm "$TEMP";
+		else
+			printf "\n%s" "The video you entered does not exist.";
+		fi
+	fi
 done
